Fix command-constructor naming collisions (#406)
This commit assigns every command node a single collision-safe generated Go constructor identifier using the `NameAllocator` during model construction. By introducing `ConstructorMethodName` on the `SubCommand` model and explicitly allocating it exactly once, we prevent the template generator from independently recalculating constructor names dynamically, which eliminates mismatches (like `NewFoo` and `NewFoo2`) and guarantees correct parent-child registration.

Added an executable `.txtar` e2e regression test to prove deterministic generation regardless of nesting levels, identically named leaf commands, and framework method collisions.
