+++ templates/testdata/cmd_positional.go.txtar
@@ -20,9 +20,7 @@
     {
       "Name": "verbose",
       "Type": "bool",
-      "FlagAliases": [
-        "v"
-      ],
+      "FlagAliases": ["v"],
       "Description": "Enable verbose output"
     }
   ]
