+++ templates/testdata/cmd_types.go.txtar
@@ -13,9 +13,7 @@
     {
       "Name": "timeout",
       "Type": "time.Duration",
-      "FlagAliases": [
-        "t"
-      ],
+      "FlagAliases": ["t"],
       "Description": "Timeout duration",
       "Default": "1s"
     },
@@ -26,10 +24,10 @@
       "Default": "3"
     },
     {
-      "Name": "config",
-      "Type": "string",
-      "Description": "Config file",
-      "Default": "config.yaml"
+        "Name": "config",
+        "Type": "string",
+        "Description": "Config file",
+        "Default": "config.yaml"
     }
   ]
 }
